refactor chart component
